refactor: rename importModule to importEsmModule, flatten folders
