+++ packages/core/src/lib/implementation/collect.ts
@@ -1,7 +1,7 @@
 import { CoreConfig, GlobalOptions, Report } from '@code-pushup/models';
 import { calcDuration } from '@code-pushup/utils';
-import { name, version } from '../../package.json';
-import { executePlugins } from './implementation/execute-plugin';
+import { name, version } from '../../../package.json';
+import { executePlugins } from './execute-plugin';
 
 export type CollectOptions = Pick<GlobalOptions, 'verbose'> &
   Pick<CoreConfig, 'plugins' | 'categories' | 'upload'>;
