+++ packages/utils/src/index.ts
@@ -9,7 +9,7 @@ export {
   objectToCliArgs,
 } from './lib/execute-process';
 export { git, latestHash } from './lib/git';
-export { importModule } from './lib/load-file';
+export { importEsmModule } from './lib/load-file';
 export { reportToMd } from './lib/report-to-md';
 export { reportToStdout } from './lib/report-to-stdout';
 export {
